fix: nutrition map state hover effect
